requirements.txt
pywin32
pystray
Pillow
customtkinter
httpx
aiohttp
pynput
qrcode[pil]
